refactor(core): split agent.ts into focused modules (#27)
Extract the 2167-line monolithic agent.ts into 7 focused modules:
- agent.ts (1332 lines) - main FrontAgent class orchestration
- helpers.ts - standalone utilities (truncateForPrompt, normalizeSearchQuery, etc.)
- rag-retrieval.ts - RAG context retrieval and query rewriting
- phase-checks.ts - phase validation (TypeScript, npm deps, code quality)
- answer-generation.ts - query answer synthesis
- memory-lifecycle.ts - cross-session memory persist/preload
- dev-server-detection.ts - dev server port detection

Public API surface unchanged: FrontAgent class and createAgent factory.
