+++ packages/core/src/agent/index.ts
@@ -1,0 +1,1 @@
+export { FrontAgent, createAgent } from './agent.js';
